Move AWS test fixture to its own module
