+++ samples/copy-with-nifi/processors/build.gradle
@@ -151,6 +151,7 @@ dependencies {
 
     testImplementation(project(":spi"))
     testImplementation(testFixtures(project(":common:util")))
+    testImplementation(testFixtures(project(":extensions:aws:aws-test")))
 }
