+++ samples/copy-with-nifi/processors/src/test/java/org/eclipse/dataspaceconnector/transfer/nifi/processors/PutS3ObjectTest.java
@@ -16,11 +16,11 @@ package org.eclipse.dataspaceconnector.transfer.nifi.processors;
 
 import com.amazonaws.services.s3.model.PutObjectResult;
 import com.amazonaws.services.s3.model.S3Object;
-import org.eclipse.dataspaceconnector.common.annotations.IntegrationTest;
-import org.eclipse.dataspaceconnector.common.testfixtures.AbstractS3Test;
-import org.eclipse.dataspaceconnector.common.testfixtures.EdcExtension;
 import org.apache.nifi.util.TestRunner;
 import org.apache.nifi.util.TestRunners;
+import org.eclipse.dataspaceconnector.aws.testfixtures.AbstractS3Test;
+import org.eclipse.dataspaceconnector.common.annotations.IntegrationTest;
+import org.eclipse.dataspaceconnector.common.testfixtures.EdcExtension;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
@@ -29,7 +29,9 @@ import java.util.HashMap;
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.eclipse.dataspaceconnector.common.testfixtures.TestUtils.*;
+import static org.eclipse.dataspaceconnector.common.testfixtures.TestUtils.SAMPLE_FILE_RESOURCE_NAME;
+import static org.eclipse.dataspaceconnector.common.testfixtures.TestUtils.getFileFromResourceName;
+import static org.eclipse.dataspaceconnector.common.testfixtures.TestUtils.getResourcePath;
 
 @IntegrationTest
 @ExtendWith(EdcExtension.class)
